feat(compression): remove redundant dtype conversion in quantization

+++ src/vidore_benchmark/evaluation/evaluate.py
@@ -54,10 +54,8 @@ def evaluate_dataset(
     if embedding_quantizer is not None:
         # NOTE: Cannot use batched quantization because the sequence lengths can be different.
         # HOTFIX: CUDA doesn't support int8 operations, so we convert the pooled embeddings back to bfloat16
-        emb_queries = [embedding_quantizer.quantize(emb_query).to(torch.bfloat16) for emb_query in emb_queries]
-        emb_documents = [
-            embedding_quantizer.quantize(emb_document).to(torch.bfloat16) for emb_document in emb_documents
-        ]
+        emb_queries = [embedding_quantizer.quantize(emb_query) for emb_query in emb_queries]
+        emb_documents = [embedding_quantizer.quantize(emb_document) for emb_document in emb_documents]
 
     if embedding_pooler is not None:
         for idx, emb_document in tqdm(enumerate(emb_documents), total=len(emb_documents), desc="Pooling embeddings..."):
